Html refactor

+++ cloudStorage.py
@@ -23,5 +23,6 @@ def deleteImage(url):
     filename = url.split("/")[-1]
 
     blob = bucket.blob("avatar/" + filename)
-
-    blob.delete()
+    print(blob.exists())
+    if blob.exists():
+        blob.delete()
